Refactor package layout for uv build
