+++ src/mcp_liquidation_map/static/favicon.ico
