+++ src/mcp_liquidation_map/static/index.html
